agents/requirements.txt
ag2
pyautogen


